+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Append dlls to final output during build. STCOR-492.
 * Avoid retrying of http requests using `useOkapiKy` hook by default. STCOR-500.
 * Expose module context via `useModules`/`withModules`/`withModule`. STCOR-502
+* Add `prefix` prop to TitleManager. STCOR-507
 
 ## [6.0.0](https://github.com/folio-org/stripes-core/tree/v6.0.0) (2020-10-06)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v5.0.2...v6.0.0)